refactor(knockdog): util -> lib 디렉터리명 변경
